Adjust the progress bar style for the funding target
