+++ types/company/profile/index.d.ts
@@ -3,3 +3,9 @@ interface DealTermElementProps {
   label: string;
   type?: "deadline";
 }
+
+interface ProgressBarProps {
+  percentage: number;
+  dayLeft: number;
+  fundingTarget: number | undefined;
+}
